test(architecture): move clinic contract tests
