+++ test/unit/contracts/clinic/clinic-permissions-middleware.test.ts
@@ -1,6 +1,6 @@
-﻿import test from "node:test";
+import test from "node:test";
 import assert from "node:assert/strict";
-import { requireClinicManagementPermission } from "../server/middlewares/clinic-permissions.ts";
+import { requireClinicManagementPermission } from "../../../../server/middlewares/clinic-permissions.ts";
 function createMockResponse() {
   return {
     statusCode: 200,
